requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
shortuuid
python-dotenv
psycopg2